The following methods (functions) are exported and available by using GLOQ.

GLOQ.LindbladODEProblem — Method

LindbladODEProblem(rho0,L::Array{ComplexF64,2},time_final::Float64;initial_type = "density")

Function provide interfaces to DifferentialEquations package.

Argument:

L: Lindblad operator

time: final time or time beining evaluated

rho0: initial condition

initial_type: specify initial value is a density matrix/a state vector

Output:

A problem object which we will feed to DifferentialEquations.jl

GLOQ.about — Method

About GLOQ.jl

GLOQ.exponential_solver — Method

exponential_solver(rho_vec0,L,t_span::Array{Float64};initial_type = "density"):

Argument:

L: the whole propagation operator

t_span: where the function will be evaluated stored in

rho0_vec: initial density/state

initial_type: decide we aer given an initial density matrix or a state vector

Output:

solutions at t_span

GLOQ.exponential_solver — Method

exponential_solver(rho_vec0,L,t_span::StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}};
                   initial_type = "density")

Argument:

L: the whole propagation operator

t_span: where the function will be evaluated stored in

rho0_vec: initial density/state

initial_type: decide we aer given an initial density matrix or a state vector

Output:

solutions at t_span

GLOQ.get_population — Method

get_population(rho_vec::Array{ComplexF64,1})

Argument:

rho_vec, density matrix in the vector form

Output:

the population in an array

GLOQ.get_population — Method

get_population(rho_vec_history::Array{ComplexF64,2})

Argument:

rhovechistory, time history for the vectorized density matrix,

i-th column corresponding to the i-th time point

Output:

The population history P which is a 2D array, P[i,j] is corresponding

to time point i and state j

GLOQ.hello_world — Method

hello_world():

A hello world function

GLOQ.make_lindblad_operator — Function

make_lindblad_operator(H,L_list,N::Int64=0)

Argument:

the Hamiltonain H and a list of Lindblad terms L_k's

N is number of states

Output:

the operator for the vectorized system

LH: corresponding to the Hamiltonian part

LD: corresponding to the Dissipation part