The following methods (functions) are exported and available by using GLOQ.

GLOQ.LindbladODEProblem — Method

LindbladODEProblem(rho_u0,L::Array{ComplexF64,2},time_final::Float64;initial_type = "density")

Function provide interfaces to DifferentialEquations package to solve the real-valued Lindblad system: $(\rho_u - i \rho_v)_t = -i (L_K+iL_S)(\rho_u-i \rho_v) + L_D(\rho_u-i \rho_v)$

Argument:

$\rho_{u0}, \rho_{v0}$: initial condition $\rho_0 = \rho_{u0} - i \rho_{v0}$

LK: real part of the Hamiltonian operator

LS: imaginary part of the Hamiltonain operator

LD: Lindblad operator

time: final time or time beining evaluated

initial_type: specify initial value is a density matrix/a state vector

Output:

A problem object which we will feed to DifferentialEquations.jl

GLOQ.LindbladODEProblemComplex — Method

LindbladODEProblemComplex(rho0,L::Array{ComplexF64,2},time_final::Float64;initial_type = "density")

Function provide interfaces to DifferentialEquations package to solve: $\rho_t = L \rho$

Argument:

L: Lindblad operator

time: final time or time beining evaluated

rho0: initial condition

initial_type: specify initial value is a density matrix/a state vector

Output:

A complex-valued problem object which we will feed to DifferentialEquations.jl

GLOQ.RamseyForwardSolve — Function

RamseyForwardSolve(rho_u0::Array{Float64},rho_v0::Array{Float64},
				   omega::Array{Float64},omega_drive:Float64,
				   gamma1::Array{Float64},gamma2::Array{Float64}
				   TC::Float64,t_dark_times::Float64,
				   N_states::Int64=0)

Argument:

rhou0,rhov0: initial states $\rho_{u_0}-i\rho_{v_0}$

omega: transition frequencies

omega_drive: driving frequency

gamma1,gamma2: determine the decay part and the dephasing part of the Lindblad operators

TC: control time of the control signal

tdarktimes: the dark time in the Ramsey experiment

N_states: number of states

Output:

rhoramseyu,rhoramseyv: density matrix at dark times, with $\rho=\rho_u-i\rho_v$

GLOQ.RotationFrameDiagonal — Method

RotationFrameDiagonal(omega::Array{Float64},omega_drive::Float64)

Argument:

omega: transition frequencies

omega_drive: driving frequency

N: number of states

Output:

Diagonal matrix diag{0,omega[1]-omr,omega[2]-omr,...,omega[N-1]-omr}

GLOQ.RotationFrameLindblad — Method

RotationFrameDiagonal(gamma1::Array{Float64},gamma2::Array{Float64})

Argument:

gamma1: determine the decay operator in the Lindblad system

gamma2: determine the dephasing operator in the Lindblad system

Output:

Ldecay: Ldecay[i][i+1] = sqrt(gamma1[i]), and zeros everywhere else

L_dephase: Diagonal{0,sqrt(gamma2[1]),...)

GLOQ.RotationFrameRamseyControl — Method

RotationFrameRamseyControl(abs_control::Float64,phase_control::Float64,N::Int64)

Argument:

abs_control: strength of the control $\Omega$

phase_control: phase of the control signal $\theta$

N: number of states

Output:

HCK,HCS: the real and imaginary part of the control operator, determined by $p(a+a^\dagger)+iq(a-a^\dagger),$ where $a$ is the lowering operator, $p=|\Omega|\cos(\theta)$ and $q=|\Omega|\sin(\theta)$

GLOQ.about — Method

About GLOQ.jl

GLOQ.convert_state_to_density — Method

convert_state_to_density(u::Array{Float64,1},v::Array{Float64,1})

Argument:

state vector: u-iv

Output:

vectorized density matrix: rhou-i rhov

GLOQ.exponential_solver — Function

exponential_solver_complex(rho_vec0,L,t_span::Array{Float64};initial_type = "density"):

Purpose:

use an exponential integrator to integrate

$\rho_t=-i(L_K+iL_S) \rho+L_D ho$

Argument:

$\rho_{u0},\rho_{v0}$: $\rho0 =$\rho{u0} - i \rho{v0}``

LK: real part of the Hamiltonian operator

LS: imaginary part of the Hamiltonain operator

LD: Lindblad operator

t_span: where the solution will be stored

initial_type: "density" vectorized density matrix, "states" states vector

Output:

rhou,rhov: solutions at tspan given by ``\rho(ti) = hou(:,i)-i hov(:,i)``

GLOQ.exponential_solver_complex — Method

exponential_solver_complex(rho_vec0,L,t_span::Array{Float64};initial_type = "density"):

Argument:

L: the whole propagation operator

t_span: where the function will be evaluated stored in

rho0_vec: initial density/state

initial_type: decide we aer given an initial density matrix or a state vector

Output:

solutions at t_span

GLOQ.exponential_solver_complex — Method

exponential_solver(rho_vec0,L,t_span::StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}};
                   initial_type = "density")

Argument:

L: the whole propagation operator

t_span: where the function will be evaluated stored in

rho0_vec: initial density/state

initial_type: decide we aer given an initial density matrix or a state vector

Output:

Complex-valued solutions at t_span

GLOQ.get_population — Method

get_population(rho_vec::Array{ComplexF64,1})

Argument:

rho_vec, density matrix in the vector form

Output:

the population in an array

GLOQ.get_population — Method

get_population(rho_vec_history::Array{ComplexF64,2})

Argument:

rhovechistory, time history for the vectorized density matrix,

i-th column corresponding to the i-th time point

Output:

The population history P which is a 2D array, P[i,j] is corresponding

to time point i and state j

GLOQ.get_population — Method

get_population(rho_vec_history::Array{Float64,2})

Argument:

rhouhistory: time history for the vectorized density matrix, j-th column corresponding to the j-th time point

rhohistory = rhouhistory - i rhov_history.

Since the density matrix should be Hermitian, it is good enough to just have the real part.

Output:

The population history P which is a 2D array, P[i,j] is corresponding

to time point i and state j

GLOQ.hello_world — Method

hello_world():

A hello world function

GLOQ.make_hamiltonian_operator — Function

make_Hamiltonian_operator(HK,HS,N::Int64=0)

Argument:

The real part and the imaginary part of the Hamiltonian: $H = H_K-i H_S$

N is number of states

Output:

the operator for the vectorized system

LK: real matrix corresponding to the real part of the Hamiltonian operator

LS: real matrix corresponding to the imaginary part of the Hamiltonian operator

GLOQ.make_lindblad_operator — Function

make_lindblad_operator(HK,HS,L_list,N::Int64=0)

Argument:

The real part and the imaginary part of the Hamiltonian: $H = H_K-i H_S$

A list of Lindblad terms $L_k$'s

N is number of states

Output:

the operator for the vectorized system

LK: real matrix corresponding to the real part of the Hamiltonian operator

LS: real matrix corresponding to the imaginary part of the Hamiltonian operator

LD: real matrix corresponding to the Lindblad operator

Final system: $\rho_t = -i(L_K+i L_S)\rho+L_D \rho$

GLOQ.make_lindblad_operator_complex — Function

make_lindblad_operator(H,L_list,N::Int64=0)

Argument:

the Hamiltonain H and a list of Lindblad terms $L_k$'s

N is number of states

Output:

the operator for the vectorized system

LH: complex matrix corresponding to the Hamiltonian operator

LD: complex matrix corresponding to the Lindblad operator